+++ src/schemathesis/_compat.py
@@ -6,7 +6,7 @@ from warnings import catch_warnings, simplefilter
 @contextmanager
 def handle_warnings() -> Generator[None, None, None]:
     try:
-        from hypothesis.errors import NonInteractiveExampleWarning
+        from hypothesis.errors import NonInteractiveExampleWarning  # pylint: disable=import-outside-toplevel
 
         with catch_warnings():
             simplefilter("ignore", NonInteractiveExampleWarning)